requirements.txt
asgiref==3.3.1
brotlipy==0.7.0
certifi==2020.12.5
Django==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==1.16.0
idna
Markdown==3.3.4
pycparser
pymongo==3.11.3
pytz==2021.1
requests
six
sqlparse==0.2.4
wincertstore==0.2
